+++ changelog.md
@@ -3,6 +3,8 @@
 ## v3.0.1
 
 1. Fixed permission issues on displaying user management buttons
+2. Fixed api config editing issues by switching to TextField
+3. Added api config editing line / column display
 
 ## v3.0.0